requirements.txt
playwright==1.42.0
playwright-stealth==1.0.6
beautifulsoup4==4.12.3
lxml==5.1.0
pandas==2.2.1
python-dotenv==1.0.1
supabase==2.4.0